# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/AirplaneTicket/search": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Search AirplaneTickets",
                "operationId": "c0ce5b7df9fa0345b7215f478205bcee",
                "parameters": [
                    {
                        "name": "arrivalTime",
                        "in": "query",
                        "description": "arrivalTime AirplaneTicket",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "id city => origin AirplaneTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "id city => destination AirplaneTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AirplaneTicket detail successfully"
                    }
                }
            }
        },
        "/api/ticket/airplane": {
            "post": {
                "tags": [
                    "Ticket Home"
                ],
                "summary": "detail AirplaneTickets",
                "operationId": "84e4913522bfd2c621cae89d0d11d990",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id AirplaneTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "AirplaneTicket detail successfully"
                    }
                }
            }
        },
        "/api/ticket/trin": {
            "post": {
                "tags": [
                    "Ticket Home"
                ],
                "summary": "detail TrainTicket",
                "operationId": "e87194074a0964bc57368f120a35c970",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "id AirplaneTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TrainTicket detail successfully"
                    }
                }
            }
        },
        "/api/TrainTicket/search": {
            "post": {
                "tags": [
                    "Search"
                ],
                "summary": "Search TrainTickets",
                "operationId": "ec60962f420dafebc801f51240f2da4e",
                "parameters": [
                    {
                        "name": "arrivalTime",
                        "in": "query",
                        "description": "arrivalTime TrainTicket",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "origin",
                        "in": "query",
                        "description": "id city => origin TrainTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "destination",
                        "in": "query",
                        "description": "id city => destination TrainTicket",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "TrainTicket detail successfully"
                    },
                    "409": {
                        "description": "TrainTicket Validate Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}